requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
jinja2==3.1.3
pydantic==2.6.4
google-generativeai==0.3.2
pdfkit==1.0.0